# docs/openapi.contracts.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/supplies/items": {
      "get": {
        "operationId": "SuppliesController_items",
        "parameters": [],
        "responses": {
          "501": {
            "description": "Contract only"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List club supply inventory",
        "tags": [
          "supplies"
        ],
        "x-implementation-status": "contract-only"
      },
      "post": {
        "operationId": "SuppliesController_createItem",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSupplyItemDto"
              }
            }
          }
        },
        "responses": {
          "501": {
            "description": "Contract only"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create supply item",
        "tags": [
          "supplies"
        ],
        "x-implementation-status": "contract-only"
      }
    },
    "/api/v1/supplies/items/low-stock": {
      "get": {
        "operationId": "SuppliesController_lowStock",
        "parameters": [],
        "responses": {
          "501": {
            "description": "Contract only"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List items at or below reorder threshold",
        "tags": [
          "supplies"
        ],
        "x-implementation-status": "contract-only"
      }
    },
    "/api/v1/supplies/items/{id}": {
      "get": {
        "operationId": "SuppliesController_item",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "501": {
            "description": "Contract only"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get supply item",
        "tags": [
          "supplies"
        ],
        "x-implementation-status": "contract-only"
      },
      "patch": {
        "operationId": "SuppliesController_updateItem",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSupplyItemDto"
              }
            }
          }
        },
        "responses": {
          "501": {
            "description": "Contract only"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update supply quantity or threshold",
        "tags": [
          "supplies"
        ],
        "x-implementation-status": "contract-only"
      },
      "delete": {
        "operationId": "SuppliesController_deleteItem",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "501": {
            "description": "Contract only"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Soft-delete supply item",
        "tags": [
          "supplies"
        ],
        "x-implementation-status": "contract-only"
      }
    },
    "/api/v1/supplies/requests": {
      "get": {
        "operationId": "SuppliesController_requests",
        "parameters": [],
        "responses": {
          "501": {
            "description": "Contract only"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List supply requests",
        "tags": [
          "supplies"
        ],
        "x-implementation-status": "contract-only"
      },
      "post": {
        "operationId": "SuppliesController_createRequest",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSupplyRequestDto"
              }
            }
          }
        },
        "responses": {
          "501": {
            "description": "Contract only"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Request supply replenishment",
        "tags": [
          "supplies"
        ],
        "x-implementation-status": "contract-only"
      }
    },
    "/api/v1/supplies/requests/{id}": {
      "get": {
        "operationId": "SuppliesController_request",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "501": {
            "description": "Contract only"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get supply request",
        "tags": [
          "supplies"
        ],
        "x-implementation-status": "contract-only"
      }
    },
    "/api/v1/supplies/requests/{id}/status": {
      "patch": {
        "operationId": "SuppliesController_updateRequest",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSupplyRequestDto"
              }
            }
          }
        },
        "responses": {
          "501": {
            "description": "Contract only"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Approve, reject or fulfill supply request",
        "tags": [
          "supplies"
        ],
        "x-implementation-status": "contract-only"
      }
    },
    "/api/v1/training-plans/{id}": {
      "get": {
        "operationId": "TrainingDetailsController_plan",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "501": {
            "description": "Contract only"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get training plan",
        "tags": [
          "training"
        ],
        "x-implementation-status": "contract-only"
      },
      "patch": {
        "operationId": "TrainingController_updatePlan",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTrainingPlanDto"
              }
            }
          }
        },
        "responses": {
          "501": {
            "description": "Contract only"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update training plan",
        "tags": [
          "training"
        ],
        "x-implementation-status": "contract-only"
      }
    },
    "/api/v1/training-plans/{id}/activate": {
      "post": {
        "operationId": "TrainingDetailsController_activatePlan",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "501": {
            "description": "Contract only"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Activate training plan",
        "tags": [
          "training"
        ],
        "x-implementation-status": "contract-only"
      }
    },
    "/api/v1/training-plans/{id}/cancel": {
      "post": {
        "operationId": "TrainingDetailsController_cancelPlan",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "501": {
            "description": "Contract only"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cancel training plan",
        "tags": [
          "training"
        ],
        "x-implementation-status": "contract-only"
      }
    },
    "/api/v1/training-plans/{id}/sessions": {
      "get": {
        "operationId": "TrainingDetailsController_sessions",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "501": {
            "description": "Contract only"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List sessions in training plan",
        "tags": [
          "training"
        ],
        "x-implementation-status": "contract-only"
      },
      "post": {
        "operationId": "TrainingController_createSession",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTrainingSessionDto"
              }
            }
          }
        },
        "responses": {
          "501": {
            "description": "Contract only"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Schedule training session",
        "tags": [
          "training"
        ],
        "x-implementation-status": "contract-only"
      }
    },
    "/api/v1/sessions/{id}": {
      "get": {
        "operationId": "TrainingDetailsController_session",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "501": {
            "description": "Contract only"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get training session",
        "tags": [
          "training"
        ],
        "x-implementation-status": "contract-only"
      },
      "patch": {
        "operationId": "TrainingDetailsController_updateSession",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTrainingSessionDto"
              }
            }
          }
        },
        "responses": {
          "501": {
            "description": "Contract only"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Reschedule or reassign training session",
        "tags": [
          "training"
        ],
        "x-implementation-status": "contract-only"
      }
    },
    "/api/v1/sessions/{id}/evaluation": {
      "get": {
        "operationId": "TrainingDetailsController_evaluation",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "501": {
            "description": "Contract only"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get session evaluation",
        "tags": [
          "training"
        ],
        "x-implementation-status": "contract-only"
      },
      "post": {
        "operationId": "TrainingController_evaluateSession",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvaluateSessionDto"
              }
            }
          }
        },
        "responses": {
          "501": {
            "description": "Contract only"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Evaluate completed session",
        "tags": [
          "training"
        ],
        "x-implementation-status": "contract-only"
      }
    },
    "/api/v1/sessions/{id}/time-trials": {
      "get": {
        "operationId": "TrainingDetailsController_timeTrials",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "501": {
            "description": "Contract only"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List session time trials",
        "tags": [
          "training"
        ],
        "x-implementation-status": "contract-only"
      },
      "post": {
        "operationId": "TrainingDetailsController_createTimeTrial",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTimeTrialDto"
              }
            }
          }
        },
        "responses": {
          "501": {
            "description": "Contract only"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Record session time trial",
        "tags": [
          "training"
        ],
        "x-implementation-status": "contract-only"
      }
    },
    "/api/v1/time-trials/{id}": {
      "get": {
        "operationId": "TrainingDetailsController_timeTrial",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "501": {
            "description": "Contract only"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get time trial result and media",
        "tags": [
          "training"
        ],
        "x-implementation-status": "contract-only"
      }
    },
...
    "/api/v1/clubs/me": {
      "get": {
        "operationId": "ClubsController_me",
        "parameters": [],
        "responses": {
          "501": {
            "description": "Contract only"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get current club",
        "tags": [
          "clubs"
        ],
        "x-implementation-status": "contract-only"
      },
      "patch": {
        "operationId": "ClubsController_update",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateClubDto"
              }
            }
          }
        },
        "responses": {
          "501": {
            "description": "Contract only"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update current club settings",
        "tags": [
          "clubs"
        ],
        "x-implementation-status": "contract-only"
      }
    },
...
  },
  "info": {
    "title": "Racehorse API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateSupplyItemDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "unit": {
            "type": "string"
          },
          "quantityOnHand": {
            "type": "number",
            "minimum": 0
          },
          "reorderThreshold": {
            "type": "number",
            "minimum": 0
          }
        },
        "required": [
          "name",
          "unit",
          "quantityOnHand",
          "reorderThreshold"
        ]
      },
      "UpdateSupplyItemDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "unit": {
            "type": "string"
          },
          "quantityOnHand": {
            "type": "number",
            "minimum": 0
          },
          "reorderThreshold": {
            "type": "number",
            "minimum": 0
          }
        }
      },
      "CreateSupplyRequestDto": {
        "type": "object",
        "properties": {
          "itemId": {
            "type": "string",
            "format": "uuid"
          },
          "quantity": {
            "type": "number",
            "minimum": 0
          },
          "note": {
            "type": "string"
          }
        },
        "required": [
          "itemId",
          "quantity"
        ]
      },
      "UpdateSupplyRequestDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "APPROVED",
              "REJECTED",
              "FULFILLED"
            ]
          }
        },
        "required": [
          "status"
        ]
      },
      "UpdateTrainingSessionDto": {
        "type": "object",
        "properties": {
          "scheduledAt": {
            "type": "string",
            "format": "date-time"
          },
          "distanceKm": {
            "type": "number",
            "minimum": 0
          },
          "intensity": {
            "type": "string"
          },
          "surface": {
            "type": "string"
          },
          "groomId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "CreateTimeTrialDto": {
        "type": "object",
        "properties": {
          "distanceMeters": {
            "type": "number",
            "minimum": 0
          },
          "durationSeconds": {
            "type": "number",
            "minimum": 0
          },
          "videoAssetId": {
            "type": "string",
            "format": "uuid"
          },
          "notes": {
            "type": "string"
          }
        },
        "required": [
          "distanceMeters",
          "durationSeconds"
        ]
      },
...
      "UpdateTrainingPlanDto": {
        "type": "object",
        "properties": {
          "phaseName": {
            "type": "string"
          },
          "goal": {
            "type": "string"
          },
          "startDate": {
            "type": "string",
            "format": "date"
          },
          "endDate": {
            "type": "string",
            "format": "date"
          }
        }
      },
      "CreateTrainingSessionDto": {
        "type": "object",
        "properties": {
          "scheduledAt": {
            "type": "string",
            "format": "date-time"
          },
          "distanceKm": {
            "type": "number",
            "minimum": 0
          },
          "intensity": {
            "type": "string"
          },
          "surface": {
            "type": "string"
          },
          "groomId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "scheduledAt",
          "distanceKm",
          "intensity"
        ]
      },
      "EvaluateSessionDto": {
        "type": "object",
        "properties": {
          "score": {
            "type": "number",
            "minimum": 1,
            "maximum": 10
          },
          "comment": {
            "type": "string"
          }
        },
        "required": [
          "score"
        ]
      },
      "UpdateClubDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
...
    }
  }
}